New main functionality: check if given symbol addresses fit
The program will now take an uneven number of arguments, first
the library to check and then at least two tuples consisting of the
name of a symbol and its address found in a program (e.g. in the GOT).
The program will then calculate the base address at which the library
was loaded and confirm that it fits to all given symbol name address
combinations.
If successful, the program will return 0 and print the load address
in hex to stdout. Otherwise, it will return != 0 and print an error
message to stderr.